refactor overall

+++ src/ecfft/isogeny.rs
@@ -9,7 +9,7 @@ pub(crate) struct Isogeny {
     b: Fp,
 }
 impl Isogeny {
-    pub(crate) fn new(depth: u32) -> Isogeny {
+    pub(crate) fn new(depth: usize) -> Isogeny {
         match depth {
             0 => Isogeny {
                 a: Fp::from_raw([
